Moved unread to app.tsx
